+++ CHANGELOG.md
@@ -2,8 +2,20 @@
 
 All notable changes to `mateusjunges/laravel-acl` will be documented in this file.
 ## 2.0.2
-- Add option to use array as parameter for `syncPermissions()` method
+#### In [UsersTrait.php](https://github.com/jungessolutions/laravel-acl/blob/master/src/Traits/UsersTrait.php):
+- Add option to use array as parameter for `syncPermissions()` method;
+- Add option to use array as parameter for `assignPermissions()` method;
+- Add option to use array as parameter for `assignGroup()` method;
+- Add option to use array as parameter for `revokePermissions()` method;
+- Add option to use array as parameter for `assignGroup()` method;
+- Add option to use array as parameter for `revokeGroup()` method;
 
+#### In [GroupsTrait.php](https://github.com/mateusjunges/laravel-acl/blob/master/src/Traits/GroupsTrait.php):
+- Add option to use array as parameter for `syncPermissions()` method;
+- Add option to use array as parameter for `assignPermissions()` method;
+- Add option to use array as parameter for `revokePermissions()` method;
+- Add option to use array as parameter for `assignUser()` method;
+- Add option to use array as parameter for `removeUser()` method;
 
 ## 2.0.1
 - Update composer.json for laravel 6.0
@@ -16,7 +28,7 @@ methods now works with non array params. You can [read the docs here](https://gi
 
 - Added eager loading for permissions checks.
 
-#### In [GroupsTrait.php]():
+#### In [GroupsTrait.php](https://github.com/mateusjunges/laravel-acl/blob/master/src/Traits/GroupsTrait.php):
 - `assignPermissions()`, `syncPermissions()`, `revokePermissions()`,
 `assignUser()`, `removeUser()`, `hasAnyPermissions()` and `hasAllPermissions()`
 methods now works with non array params. You can [read the docs here](https://github.com/jungessolutions/laravel-acl#checking-for-permissions);